Move datacollection wrappers to seperate files
