+++ src/DataCollectors/TraceDataCollector/DataCollection/Interfaces/IDataCollectionAgentContext.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Microsoft.VisualStudio.TraceCollector
+{
+    using TestPlatform.ObjectModel.DataCollection;
+
+    internal interface IDataCollectionAgentContext
+    {
+        DataCollectionContext SessionDataCollectionContext { get; }
+    }
+}
